Ability for KRS to collect payment as part of recovery transaction
Summary:
This change adds an example for KRS operator to code a fee percentage/minimum recovery fee into the recovery process.
The resultant transaction will have 2 outputs - an output to the customer's destination address, and one to the KRS provider's payment address
The constructed TX should can/still be inspected by the KRS operator since this construction is client side.

Reviewers: ben, arik

Reviewed By: arik

Differential Revision: https://phabricator.bitgo.com/D1720